requirements.txt
crewai==0.28.0  # Downgraded to support pydantic v1
streamlit==1.38.0
google-generativeai==0.3.1
python-dotenv==1.0.0
pandas==2.2.3
langchain==0.1.10
langchain-google-genai== 0.0.9
pydantic==2.4.2